requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0

# Federated Learning
flwr>=1.5.0

# Privacy
opacus>=1.4.0

# Explainability
captum>=0.6.0
matplotlib>=3.7.0
opencv-python>=4.8.0

# Dashboard
streamlit>=1.25.0
plotly>=5.15.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
pillow>=10.0.0

# Experiment Tracking
wandb>=0.15.0

# Testing
pytest>=7.4.0